requirements.txt
beautifulsoup4==4.8.1
billboard.py==5.3.0
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
idna==2.8
joblib==0.14.1
kiwisolver==1.1.0
matplotlib==3.1.2
numpy==1.17.4
pandas==0.25.3
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
scikit-learn==0.22
scipy==1.3.3
six==1.13.0
sklearn==0.0
soupsieve==1.9.5
tqdm==4.40.2
urllib3==1.25.7
xgboost==0.90
